+++ cloudflare/src/handle.rs
@@ -10,10 +10,21 @@ use tailcall::config::Config;
 use tailcall::http::{handle_request, AppContext};
 use tailcall::io::EnvIO;
 
+use crate::http::{to_request, to_response};
 use crate::{init_env, init_file, init_http};
 
 lazy_static! {
   static ref APP_CTX: RwLock<Option<Arc<AppContext>>> = RwLock::new(None);
+}
+///
+/// The main fetch handler that handles requests on cloudflare
+///
+pub async fn fetch(req: worker::Request, env: worker::Env, _: worker::Context) -> anyhow::Result<worker::Response> {
+  let env = Rc::new(env);
+  log::debug!("Execution starting");
+  let app_ctx = get_app_ctx(env).await?;
+  let resp = handle_request::<GraphQLRequest>(to_request(req).await?, app_ctx).await?;
+  Ok(to_response(resp).await?)
 }
 
 ///
@@ -28,19 +39,11 @@ async fn get_config(env_io: &impl EnvIO, env: Rc<worker::Env>) -> anyhow::Result
   Ok(config)
 }
 
-pub async fn fetch(req: worker::Request, env: worker::Env, _: worker::Context) -> anyhow::Result<worker::Response> {
-  let env = Rc::new(env);
-  log::debug!("Execution starting");
-  let app_ctx = init(env).await?;
-  let resp = handle_request::<GraphQLRequest>(to_request(req).await?, app_ctx).await?;
-  Ok(to_response(resp).await?)
-}
-
 ///
 /// Initializes the worker once and caches the app context
 /// for future requests.
 ///
-async fn init(env: Rc<worker::Env>) -> anyhow::Result<Arc<AppContext>> {
+async fn get_app_ctx(env: Rc<worker::Env>) -> anyhow::Result<Arc<AppContext>> {
   // Read context from cache
   if let Some(app_ctx) = read_app_ctx() {
     Ok(app_ctx.clone())
@@ -60,52 +63,4 @@ async fn init(env: Rc<worker::Env>) -> anyhow::Result<Arc<AppContext>> {
 
 fn read_app_ctx() -> Option<Arc<AppContext>> {
   APP_CTX.read().unwrap().clone()
-}
-
-pub async fn to_response(response: hyper::Response<hyper::Body>) -> anyhow::Result<worker::Response> {
-  let status = response.status().as_u16();
-  let headers = response.headers().clone();
-  let bytes = hyper::body::to_bytes(response).await?;
-  let body = worker::ResponseBody::Body(bytes.to_vec());
-  let mut w_response = worker::Response::from_body(body).map_err(to_anyhow)?;
-  w_response = w_response.with_status(status);
-  let mut_headers = w_response.headers_mut();
-  for (name, value) in headers.iter() {
-    let value = String::from_utf8(value.as_bytes().to_vec())?;
-    mut_headers.append(name.as_str(), &value).map_err(to_anyhow)?;
-  }
-
-  Ok(w_response)
-}
-
-fn to_method(method: worker::Method) -> anyhow::Result<hyper::Method> {
-  let method = &*method.to_string().to_uppercase();
-  match method {
-    "GET" => Ok(hyper::Method::GET),
-    "POST" => Ok(hyper::Method::POST),
-    "PUT" => Ok(hyper::Method::PUT),
-    "DELETE" => Ok(hyper::Method::DELETE),
-    "HEAD" => Ok(hyper::Method::HEAD),
-    "OPTIONS" => Ok(hyper::Method::OPTIONS),
-    "PATCH" => Ok(hyper::Method::PATCH),
-    "CONNECT" => Ok(hyper::Method::CONNECT),
-    "TRACE" => Ok(hyper::Method::TRACE),
-    method => Err(anyhow!("Unsupported HTTP method: {}", method)),
-  }
-}
-
-pub async fn to_request(mut req: worker::Request) -> anyhow::Result<hyper::Request<hyper::Body>> {
-  let body = req.text().await.map_err(to_anyhow)?;
-  let method = req.method();
-  let uri = req.url().map_err(to_anyhow)?.as_str().to_string();
-  let headers = req.headers();
-  let mut builder = hyper::Request::builder().method(to_method(method)?).uri(uri);
-  for (k, v) in headers {
-    builder = builder.header(k, v);
-  }
-  Ok(builder.body(hyper::body::Body::from(body))?)
-}
-
-fn to_anyhow<T: std::fmt::Display>(e: T) -> anyhow::Error {
-  anyhow!("{}", e)
 }
